% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloraIberica-package.R
\docType{package}
\name{FloraIberica-package}
\alias{FloraIberica}
\alias{FloraIberica-package}
\title{FloraIberica: Taxonomic and distribution data for all the vascular plants present in the Iberian Peninsula and/or Balearic Islands}
\description{
Check which plant taxa are present and/or endemic, and get detailed (10x10 km) distribution maps for all the vascular plants present in the Iberian Peninsula and Balearic Islands, from genus to subspecies level, based on the AFLIBER database (https://doi.org/10.1111/geb.13363).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/FloraIberica}
  \item Report bugs at \url{https://github.com/Pakillo/FloraIberica/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodríguez-Sánchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
