% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checklist.R
\name{get_checklist}
\alias{get_checklist}
\title{Get a zonal checklist}
\usage{
get_checklist(zone = NULL, sf = FALSE)
}
\arguments{
\item{zone}{A vector of two numbers giving the point longitude and latitude (in that order),
an \code{sf} object with a single point coordinate, or a polygon \code{sf} object.}

\item{sf}{Logical. If FALSE (default) return a dataframe with the checklist.
If TRUE, return an \code{sf} object with the coordinates where each taxa is found.}
}
\value{
A dataframe or sf object
}
\description{
Get a checklist of the vascular plants growing near a given point or polygon.
}
\note{
As the original data (AFLIBER database) have 10-km resolution, the
resulting checklist may include taxa present within 10 km distance of the point or
polygon.
}
\examples{
sitio <- c(-5, 40)
head(get_checklist(sitio))

}
