% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution.R
\name{get_distribution}
\alias{get_distribution}
\title{Get the distribution of one or more plant taxa}
\usage{
get_distribution(genus = NULL, species = NULL, subspecies = NULL, sf = TRUE)
}
\arguments{
\item{genus}{character. Required. One or more genera to check for presence.}

\item{species}{character. Optional. One or more species names to check for presence.
If >1 species, they must all belong to the same genus.}

\item{subspecies}{character. Optional. One or more subspecies names to check for presence.
If >1 subspecies, they must all belong to the same species.}

\item{sf}{Logical. Return a spatial (sf) or a plain dataframe?}
}
\value{
An sf object if sf = TRUE, a plain dataframe otherwise.
}
\description{
Get the distribution of one or more plant taxa
}
\examples{
abies <- get_distribution("Abies")
abies
unique(abies$Species)  # including all species in the genus

pinsapo <- get_distribution("Abies", "pinsapo")
pinsapo
}
