% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IberianPeninsula.R
\docType{data}
\name{IberianPeninsula}
\alias{IberianPeninsula}
\title{Iberian Peninsula contour}
\format{
\subsection{\code{IberianPeninsula}}{

A sfc_MULTIPOLYGON object with the contour of Portugal, Peninsular Spain and Balearic Islands
}
}
\source{
\url{https://www.naturalearthdata.com/}
}
\usage{
IberianPeninsula
}
\description{
Iberian Peninsula contour
}
\keyword{datasets}
