% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_endemic.R
\name{is_endemic}
\alias{is_endemic}
\title{Is taxon endemic?}
\usage{
is_endemic(genus = NULL, species = NULL, subspecies = NULL, gbif.id = NULL)
}
\arguments{
\item{genus}{character. Required. One or more genera to check for presence.}

\item{species}{character. Optional. One or more species names to check for presence.
If >1 species, they must all belong to the same genus.}

\item{subspecies}{character. Optional. One or more subspecies names to check for presence.
If >1 subspecies, they must all belong to the same species.}

\item{gbif.id}{character. Optional.}
}
\value{
A logical vector
}
\description{
Is taxon endemic of the Iberian Peninsula and/or Balearic Islands?
That is, not present naturally elsewhere in the world.
}
\details{
For a genus to be considered endemic, all its species must be endemic.
Likewise, for a species to be considered endemic, all its subspecies must be endemic.
}
\examples{
is_endemic("Laurus")
is_endemic("Aconitum")
is_endemic("Aconitum", "variegatum")
is_endemic("Aconitum", "napellus")
is_endemic("Aconitum", "napellus", c("vulgare", "castellanum", "lusitanicum"))
}
