% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_present.R
\name{is_present}
\alias{is_present}
\title{Is taxon present in the Iberian Peninsula and/or Balearic Islands?}
\usage{
is_present(genus = NULL, species = NULL, subspecies = NULL, gbif.id = NULL)
}
\arguments{
\item{genus}{character. One or more genera to check for presence. Required,
unless \code{gbif.id} is provided.}

\item{species}{character. Optional. One or more species names to check for presence.
The length of \code{genus} must equal that of \code{species}, unless length(genus) == 1,
in which case it will be assumed that all species belong to that same genus.}

\item{subspecies}{character. Optional. One or more subspecies names to check for presence.
The length of \code{species} must equal that of \code{subspecies}, unless length(species) == 1,
in which case it will be assumed that all subspecies belong to that same species.}

\item{gbif.id}{character. Optional.}
}
\value{
A logical vector
}
\description{
Is taxon present in the Iberian Peninsula and/or Balearic Islands?
}
\examples{
is_present("Laurus")
is_present("Laurus", "nobilis")
is_present("Laurus", "azorica")
is_present("Laurus", c("nobilis", "azorica"))
is_present("Laurus", c("nobilis", "communis"))
is_present("Laurus", "nobilis", subspecies = "excelsa")
is_present(gbif.id = "3034015")
}
