% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_distribution.R
\name{map_distribution}
\alias{map_distribution}
\title{Map taxa distributions}
\usage{
map_distribution(
  distrib.sf = NULL,
  genus = NULL,
  species = NULL,
  subspecies = NULL,
  facet = FALSE,
  colour = "medium sea green",
  include.name = TRUE,
  ...
)
}
\arguments{
\item{distrib.sf}{An sf object as returned by \code{\link[=get_distribution]{get_distribution()}}.}

\item{genus}{character. Required. One or more genera to check for presence.}

\item{species}{character. Optional. One or more species names to check for presence.
If >1 species, they must all belong to the same genus.}

\item{subspecies}{character. Optional. One or more subspecies names to check for presence.
If >1 subspecies, they must all belong to the same species.}

\item{facet}{Logical. For multiple taxa, make a single map with all taxa together,
or make a multipanel (facetted) figure with one panel per taxa?}

\item{colour}{character. When there is >1 taxon, only used if facet = TRUE.}

\item{include.name}{Logical. When there is a single taxon to map, use taxon
name as title?}

\item{...}{Further params to be passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} if facet = TRUE,
or to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} if facet = FALSE.}
}
\value{
A map and ggplot2 object
}
\description{
Provide either an sf object or genus, species, subspecies...
}
\examples{
laurus.sf <- get_distribution("Laurus", "nobilis")
map_distribution(laurus.sf)

map_distribution(genus = "Laurus", species = "nobilis")

abies.sf <- get_distribution("Abies")
map_distribution(abies.sf)
map_distribution(abies.sf, facet = TRUE, ncol = 1)
}
