% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.R
\docType{data}
\name{Taxa}
\alias{Taxa}
\title{Plant taxa}
\format{
\subsection{\code{Taxa}}{

A data frame with 6,456 rows and 11 columns:
\describe{
\item{Taxon}{Full taxon name}
\item{Scientific_Name}{Scientific name}
\item{Endemic}{Endemic in the Iberian Peninsula and/or Balearic Islands (TRUE/FALSE)}
\item{Genus}{genus}
\item{Species}{species}
\item{Subspecies}{subspecies}
\item{Class}{class}
\item{Order}{order}
\item{Family}{family}
\item{GBIF_id}{GBIF unique ID for that taxon}
\item{POW_Name}{Name of the taxon in Plants of the World database}
}
}
}
\source{
\url{https://doi.org/10.1111/geb.13363}
}
\usage{
Taxa
}
\description{
Plant taxa
}
\keyword{datasets}
