% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\docType{data}
\name{Distributions}
\alias{Distributions}
\title{Taxa distributions}
\format{
\subsection{\code{Distributions}}{

A data frame with 1,824,549 rows and 6 columns:
\describe{
\item{Genus}{genus}
\item{Species}{species}
\item{Subspecies}{subspecies}
\item{UTM.cell}{Name of the 10x10 km UTM.cell}
\item{lng}{longitude in decimal degrees}
\item{lat}{latitude in decimal degrees}
}
}
}
\source{
\url{https://doi.org/10.1111/geb.13363}
}
\usage{
Distributions
}
\description{
Taxa distributions
}
\keyword{datasets}
