% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_present.R
\name{is_present}
\alias{is_present}
\title{Is taxon present?}
\usage{
is_present(genus = NULL, species = NULL, subspecies = NULL, gbif.id = NULL)
}
\arguments{
\item{genus}{character. Required. One or more genera to check for presence.}

\item{species}{character. Optional. One or more species names to check for presence.
If >1 species, they must all belong to the same genus.}

\item{subspecies}{character. Optional. One or more subspecies names to check for presence.
If >1 subspecies, they must all belong to the same species.}

\item{gbif.id}{character. Optional.}
}
\value{
A logical vector
}
\description{
Is taxon present in the Iberian Peninsula and/or Balearic Islands?
}
\examples{
is_present("Laurus")
is_present("Laurus", "nobilis")
is_present("Laurus", "azorica")
is_present("Laurus", c("nobilis", "azorica"))
is_present(gbif.id = "3034015")
}
