% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_distribution.R
\name{map_distribution}
\alias{map_distribution}
\title{Map taxa distributions}
\usage{
map_distribution(
  distrib.sf = NULL,
  genus = NULL,
  species = NULL,
  subspecies = NULL,
  taxo.level = "species",
  facet = FALSE,
  colour = "medium sea green",
  include.name = TRUE,
  ...
)
}
\arguments{
\item{distrib.sf}{An sf object as returned by \code{\link[=get_distribution]{get_distribution()}}.}

\item{genus}{character. One or more genera to check for presence. Required,
unless \code{gbif.id} is provided.}

\item{species}{character. Optional. One or more species names to check for presence.
The length of \code{genus} must equal that of \code{species}, unless length(genus) == 1,
in which case it will be assumed that all species belong to that same genus.}

\item{subspecies}{character. Optional. One or more subspecies names to check for presence.
The length of \code{species} must equal that of \code{subspecies}, unless length(species) == 1,
in which case it will be assumed that all subspecies belong to that same species.}

\item{taxo.level}{character Taxonomic level to show in the map. Either genus,
species (default) or subspecies.}

\item{facet}{Logical. For multiple taxa, make a single map with all taxa together,
or make a multipanel (facetted) figure with one panel per taxa?}

\item{colour}{character. When there is >1 taxon, only used if facet = TRUE.}

\item{include.name}{Logical. When there is a single taxon to map, use taxon
name as title?}

\item{...}{Further params to be passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} if facet = TRUE,
or to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} if facet = FALSE.}
}
\value{
A map and ggplot2 object
}
\description{
Get a map with the distribution of one or more taxa. Must provide either
an \code{sf} object (as returned by \code{\link[=get_distribution]{get_distribution()}}) or genus, species, subspecies...
}
\examples{
laurus <- get_distribution("Laurus", "nobilis")
map_distribution(laurus)

map_distribution(genus = "Laurus", species = "nobilis")

abies <- get_distribution("Abies")
map_distribution(abies)
map_distribution(abies, facet = TRUE, ncol = 1)
map_distribution(abies, taxo.level = "genus")
}
